+++ R/glmerb_posterior_mode.R
@@ -51,7 +51,7 @@
 #'   \code{iterations}, and \code{delta}.
 #'
 #' @seealso \code{\link{lmerb_posterior_mean}}, \code{\link{rglmb}},
-#'   \code{\link{two_block_rNormal_reg_v2}}, \code{\link{build_mu_all}}
+#'   \code{\link{two_block_rNormal_reg}}, \code{\link{build_mu_all}}
 #' @export
 glmerb_posterior_mode <- function(design,
                                   family = gaussian(),